requirements.txt
Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
langchain==0.0.335
langchain-openai==0.0.2
cohere==4.32
langchain-cohere==0.0.2
langchain-community==0.0.13
google-generativeai==0.3.1
langchain-google-genai==0.0.7
pypdf==3.17.1
docx2txt==0.8
langchain-chroma==0.0.2
chromadb==0.4.22
matplotlib==3.8.0
plotly==5.18.0
pandas==2.1.1
flask-socketio==5.3.6
pillow==10.1.0
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-image==0.21.0
SpeechRecognition==3.10.0
pyttsx3==2.90
redis==5.0.1
sse-starlette==1.6.5
boto3==1.34.33
flask-bcrypt==1.0.1
flask-login==0.6.3 